FastAPI/requirements.txt
absl-py==1.4.0
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.101
botocore==1.29.101
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fastapi==0.95.0
fonttools==4.39.3
google-auth==2.17.1
google-auth-oauthlib==0.4.6
grpcio==1.53.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.12.0
jedi==0.18.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.13
uvicorn==0.21.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
